refactor(OutputRedirection): move code to a separate file
